refactor: rename 'components' to 'view'
