+++ src/App.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { DarkTheme } from 'react-native-paper'
 import { enableScreens } from 'react-native-screens'
 import { AppContainer } from './navigators/App'
-import { StatusBar } from './components/atoms/StatusBar'
+import { StatusBar } from './view/atoms/StatusBar'
 import { AuthBlocProvider } from './hooks/useAuthBloc'
 import { ThemeBlocProvider } from './hooks/useThemeBloc'
 import { PostSendBlocProvider } from './hooks/usePostSendBloc'
