+++ src/navigators/App.tsx
@@ -1,9 +1,9 @@
 import { createAppContainer, createStackNavigator, createSwitchNavigator } from 'react-navigation'
 import { MainTab } from './Main'
 import { AuthStack } from './Auth'
-import { LoadingScreen } from '../components/pages/LoadingScreen'
-import { PostModalScreen } from '../components/pages/PostModalScreen'
-import { FileModalScreen } from '../components/pages/FileModalScreen'
+import { LoadingScreen } from '../view/pages/LoadingScreen'
+import { PostModalScreen } from '../view/pages/PostModalScreen'
+import { FileModalScreen } from '../view/pages/FileModalScreen'
 
 const AppStack = createStackNavigator(
   {
