+++ src/view/atoms/StatusBar.tsx
