+++ src/view/hocs/withNavigationOption.ts
