+++ src/view/templates/AuthorizeWithCodeScreenView.tsx
