+++ src/view/templates/FileModalScreenView/story.js
