+++ src/view/templates/HomeScreenView.tsx
