+++ src/view/templates/PostModalScreenView/story.js
