+++ src/view/templates/SignInScreenView.tsx
